requirements.txt
fastapi
pillow
uvicorn
pydantic
python-dotenv
google-genai